api/pyproject.toml
[tool.poetry]
name = "api"
version = "0.1.0"
description = "Blog Api written using FastApi"
authors = ["Soroush Alinia <soroushalinia.wm@gmail.com>"]

[tool.poetry.dependencies]
python = "^3.9"
uvicorn = "^0.17.5"
asyncpg = "^0.25.0"
SQLAlchemy = "^1.4.31"
psycopg2-binary = "^2.9.3"
redis = "^4.1.4"
fastapi-mail = "^1.0.4"
python-multipart = "^0.0.5"
bcrypt = "^3.2.0"
passlib = {extras = ["bcrypt"], version = "^1.7.4"}
python-jose = "^3.3.0"
alembic = "^1.7.6"

[tool.poetry.dev-dependencies]
pytest = "^5.2"
requests = "^2.27.1"
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
